requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-telegram-bot==21.4
python-dotenv==1.0.1
